made digital core block reorder rather than retime
