--- tests/cpu_block_tests/retimer/test_retimer.py
@@ -1,37 +1,0 @@
-# general imports
-import numpy as np
-from pathlib import Path
-
-# DragonPHY imports
-from dragonphy import *
-
-THIS_DIR = Path(__file__).parent.resolve()
-BUILD_DIR = THIS_DIR / 'build'
-
-def test_sim(dump_waveforms):
-    deps = get_deps_cpu_sim(impl_file=THIS_DIR / 'test.sv')
-    print(deps)
-
-    def qwrap(s):
-        return f'"{s}"'
-
-    defines = {
-        'OUT_TXT' : qwrap(BUILD_DIR / 'ti_adc.txt')
-    }
-
-    DragonTester(
-        ext_srcs=deps,
-        directory=BUILD_DIR,
-        defines=defines,
-        dump_waveforms=dump_waveforms
-    ).run()
-
-    y = np.loadtxt(BUILD_DIR / 'ti_adc.txt', dtype=int, delimiter=',')
-    y = y.flatten()
-
-    y_ = np.arange(256)
-
-    assert(np.sum(y-y_) == 0)
-
-if __name__ == "__main__":
-    test_sim()
